cmake: install the application icon
Copy part of the KDE4Macros.cmake module from kdelibs (just the parts
of it dealing with icons installation) as KDE4Macros-Icecream.cmake
(to avoid clash with the kdelibs-provided version), along with its
license.

Make use of kde4_install_icons to properly install the icons in the
global icon theme.