Improve how docbook2man is searched for.
On ArchLinux, the docbook2X package installs docbook2man. The only way
to force this was to set -D_docbook_to_man_executable on the command
line, which is setting an internal variable.

The new approach is to only pass DOCBOOK_TO_MAN_EXECUTABLE to
find_program (which allows it to be overridden with
-DDOCBOOK_TO_MAN_EXECUTABLE), or to check for each executable name in
turn, using the first one that comes from Docbook2X.

This also makes FindDocbook2X use find_package_handle_standard_args, for
consistent messages etc.