Fix: scrollview not resized at startup + Improvement
With 5 hosts capable of handling up to 16 jobs, scrollview
is not sized correctly.

Improvement: displays average time for each submitted jobs. Target goal
is to be able to compare build times for large scale project with more
or less parallelism in order to choose the best number of simultaneous
builds / building machines.