Update views to handle node remove and node add events
Switching between views clear all the data. For example,
switching back to flow or summary view causes nodes to
disappear. Also, if a icecc host is disconnected, it's
still displayed in the view.

This is an attempt to workaround the above issues.
Scan host map when the view is created so that the hosts
are not lost when switching between views.

When host is disconnected, remove it from the host map
and update the data displayed in the view accordingly.

This involves overriding setMonitor() and addint
createKnownHosts() callback to the view classes.
With the list view we have to recreate known hosts
when a host node is removed since the rowId -> hostId
map gets messed up when removing a host.

The nodeRemoved signal is emitted when the node
is already removed and its HostInfo is deleted.

Signed-off-by: Valentine Barshak <gvaxon@gmail.com>