Fix usage count value in the flow table
When switching back to flow table form some other view
we may get negative job count value for a node. This happens
because the job counter is not saved in the HostInfo structure.
it is recreated each time we switch to the flow table view.
This adds active job counter to the HostInfo structure and
updates it whenever a job is started or finished.
This counter is displayed in the flow view.
While at it, do not display the host disabled if the active
job counter is greater than 0 even if the speed is zero.

Signed-off-by: Valentine Barshak <gvaxon@gmail.com>